Add !reload command, !pet command, fix cpu issues, added potato mode, updated websocket connectivity and frequency of saves and much more

+++ Assets/Scripts/Network/PacketHandlers/HighestSkill.cs
@@ -67,7 +67,7 @@ public class HighestSkill : PacketHandler<HighestSkillRequest>
             }
             else if (!anySkill)
             {
-                var secondary = player.GetSkill((int)skill);
+                var secondary = player.GetSkill(skill);
                 if (highest == null || highest.Experience < secondary.Experience)
                 {
                     p = player;
